backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
strands-agents[openai]==1.27.0
openai==1.82.0
supabase==2.13.0
pdfplumber==0.11.4
tiktoken==0.9.0
python-dotenv==1.1.0
matplotlib==3.10.1
pydantic==2.11.1
pytest==8.3.5
